use Map Style Provider for inset map
refactors map style provider into its own component
